correcion de microservicos 4,5 y 13

+++ m13-division-service/src/main/java/com/example/m13_division_service/repository/DivisionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DivisionRepository extends JpaRepository<Division, Integer> {
+public interface DivisionRepository extends JpaRepository<Division, Long> { // Cambiado Integer por Long
     boolean existsByNombre(String nombre);
     boolean existsByAbreviatura(String abreviatura);
 }
